apps/flashcards/requirements-edge-tts.txt
# Edge TTS Server Requirements
# Completely FREE - No API keys needed!

# Core dependencies
Flask==3.0.0
flask-cors==4.0.0
edge-tts==6.1.9

# Optional: for better async performance
aiohttp==3.9.1
